% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_data.R
\name{append_data}
\alias{append_data}
\title{Append data subsets to job list}
\usage{
append_data(input, data)
}
\arguments{
\item{input}{job list}

\item{data}{data you will subset}

\item{fold_id}{list item that corresponds to what to divide the test and training data by}
}
\value{
list of jobs with subsetted data appended
}
\description{
Append data subsets to job list
}
