% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_models.R
\name{compile_models}
\alias{compile_models}
\title{Select models to compile and compile them}
\usage{
compile_models(jobs_list, mass, initial_mass, time, group_id)
}
\arguments{
\item{jobs_list}{job list}

\item{mass}{column name in data referring to logged mass remaining}

\item{initial_mass}{columne name in data referring to logged initial mass}

\item{time}{column name in data referring to time in years}

\item{group_id}{column name in data referring to random effects cluster}
}
\value{
saves compiled models in previously specified output folder
}
\description{
Select models to compile and compile them
}
