% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jobs.R
\name{create_jobs}
\alias{create_jobs}
\title{Create job list of all model iterations to be run}
\usage{
create_jobs(model_structures, data, folder, cv_cluster = NULL,
  fixed_effects_list = NULL)
}
\arguments{
\item{model_structures}{dataframe listing model types and specifications}

\item{data}{dataframe containing data for models}

\item{folder}{folder path to save output}

\item{cv_cluster}{optional columns name to specify cv clusters}

\item{fixed_effects_list}{optional list of fixed effects - corresponding to column names in data}
}
\value{
list of all model iterations
}
\description{
Create job list of all model iterations to be run
}
\examples{
\dontrun{
traits <- c('N', 'C', 'SLA', 'LDMC', 'HC', 'CL', 'LG')
model_df <- data.frame(model_type = c('w', 'w', 'ne'),
                       fixed_effects = c(TRUE, FALSE, FALSE),
                       random_effects = c(FALSE, TRUE, FALSE),
                       cv = c(TRUE, TRUE, TRUE))
jobs_list <- create_jobs(model_structures = model_df,
                         data = df,
                         folder = 'output/stan/',
                         cv_cluster = 'species_code',
                         fixed_effects_list = traits)
}

}
