% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_output.R
\name{extract_output}
\alias{extract_output}
\title{Run single model, using the compiled files}
\usage{
extract_output(job, mass, initial_mass, time, group_id)
}
\arguments{
\item{job}{job}

\item{mass}{column name in data referring to logged mass remaining}

\item{initial_mass}{columne name in data referring to logged initial mass}

\item{time}{column name in data referring to time in years}

\item{group_id}{column name in data referring to random effects cluster}
}
\value{
list for single model with model type, iteration, neg log likelihood output, predicted and real for test data, and diagnostics
}
\description{
Run single model, using the compiled files
}
