% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Expand each cv model to have separate model run per cluster}
\usage{
run_models(jobs_list, mass, initial_mass, time, group_id)
}
\arguments{
\item{jobs_list}{job list}

\item{mass}{column name in data referring to logged mass remaining}

\item{initial_mass}{columne name in data referring to logged initial mass}

\item{time}{column name in data referring to time in years}

\item{group_id}{column name in data referring to random effects cluster}
}
\value{
list with model type, iteration, neg log likelihood output, predicted and real for test data, and diagnostics
}
\description{
Expand each cv model to have separate model run per cluster
}
