% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stan_models.R
\name{run_stan_models}
\alias{run_stan_models}
\title{Run stan models from job list}
\usage{
run_stan_models(job, mass, initial_mass, time, group_id = NULL,
  compile_model = TRUE)
}
\arguments{
\item{job}{job from job list}

\item{mass}{column name in data referring to logged mass remaining}

\item{initial_mass}{columne name in data referring to logged initial mass}

\item{time}{column name in data referring to time in years}

\item{group_id}{column name in data referring to random effects cluster}

\item{compile_model}{whether to fully compile model or not}
}
\value{
stan output, and saved .rds file
}
\description{
Run stan models from job list
}
