% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_data.R
\name{append_data}
\alias{append_data}
\title{Append data subsets to job list}
\usage{
append_data(input, data, cv_cluster)
}
\arguments{
\item{input}{job list item}

\item{data}{dataframe containing data for models}

\item{cv_cluster}{optional columns name to specify cv clusters}
}
\value{
list of jobs with subsetted data appended
}
\description{
Append data subsets to job list
}
