% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jobs.R
\name{create_jobs}
\alias{create_jobs}
\title{Create design matrix of unique model combinations}
\usage{
create_jobs(model_structures, data, folder, cv_cluster = NULL,
  fixed_effects_list = NULL)
}
\arguments{
\item{model_structures}{dataframe listing model types and specifications}

\item{data}{dataframe containing data for models}

\item{folder}{folder to save output}

\item{cv_cluster}{optional specification for grouping of cv clusters}

\item{trait_list}{a list of traits}
}
\value{
dataframe of all model structures
}
\description{
Create design matrix of unique model combinations
}
\examples{
traits <- c('N', 'C', 'SLA', 'LDMC', 'HC', 'CL', 'LG')
model_df <- data.frame(model_type = c('w', 'w', 'ne'),
                       fixed_effects = c(TRUE, FALSE, FALSE),
                       random_effects = c(FALSE, TRUE, FALSE),
                       cv = c(TRUE, TRUE, TRUE))
jobs_list <- create_jobs(model_df, df,
                         folder = 'output/stan/',
                         cv_cluster = 'species_code',
                         fixed_effects_list = traits)

}
