% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Select models to compile and compile them}
\usage{
evaluate_models(output_list, predR2_path)
}
\arguments{
\item{output_list}{list of output of run_models}

\item{predR2_path}{file path for where to save pred v. real test data plot}
}
\value{
list with deviance per model type and not converged models
}
\description{
Select models to compile and compile them
}
