% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_models.R
\name{expand_models}
\alias{expand_models}
\title{Expand each cv model to have separate model run per cluster}
\usage{
expand_models(input, clusters, folder)
}
\arguments{
\item{input}{job list item}

\item{clusters}{unique list of cv cluster id's}

\item{folder}{file path for output}
}
\value{
list of jobs additional models added
}
\description{
Expand each cv model to have separate model run per cluster
}
