% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_output.R
\name{extract_output}
\alias{extract_output}
\title{Expand each cv model to have separate model run per cluster}
\usage{
extract_output(job)
}
\arguments{
\item{job}{job list item}
}
\value{
list with model type, iteration, neg log likelihood output, and predicted and real for test data
}
\description{
Expand each cv model to have separate model run per cluster
}
